fix: log urls which get rejected while riderection after checking whitelist
